reorganized utils into separate files under a utils folder
