+++ webapp/src/components/jira_issue_selector/jira_issue_selector.jsx
@@ -7,7 +7,7 @@ import PropTypes from 'prop-types';
 import debounce from 'debounce-promise';
 import AsyncSelect from 'react-select/lib/Async';
 
-import {getStyleForReactSelect} from '../utils';
+import {getStyleForReactSelect} from 'utils/styles';
 
 const searchDefaults = 'ORDER BY updated DESC';
 const searchDebounceDelay = 400;
