+++ webapp/src/components/post_menu_actions/attach_comment_to_issue/index.js
@@ -10,7 +10,7 @@ import {isSystemMessage} from 'mattermost-redux/utils/post_utils';
 import {openAttachCommentToIssueModal} from 'actions';
 
 import {getCurrentUserLocale, isUserConnected, isInstanceInstalled} from 'selectors';
-import {isCombinedUserActivityPost} from '../../utils';
+import {isCombinedUserActivityPost} from 'utils/posts';
 
 import AttachCommentToIssuePostMenuAction from './attach_comment_to_issue';
