+++ webapp/src/utils/styles.js
@@ -64,9 +64,3 @@ export const getStyleForReactSelect = (theme) => {
         }),
     };
 };
-
-// TODO: change to use the isCombinedUserActivityPost in 'mattermost-redux/utils/post_list' when we upgrade to 5.12
-export const isCombinedUserActivityPost = (id) => {
-    return (/^user-activity-(?:[^_]+_)*[^_]+$/).test(id);
-};
-
